Move store brand info to shared module
